refactor: move invoice generation code to payment-proof folder
