+++ src/app/constants/time.ts
@@ -1,0 +1,1 @@
+export const dayInSeconds = 24 * 60 * 60
